+++ README.md
@@ -1,1 +1,9 @@
 # gitlab-cicd-dashboard
+
+Usage:
+
+$export API_TOKEN=your token
+
+$node app.js
+
+Browser: http://localhost:8081/groups/ group ID /jobs
